refactor models to seperate files
