+++ skill-manager/skill_manager/models/__init__.py
@@ -1,0 +1,2 @@
+from skill_manager.models.prediction import Prediction
+from skill_manager.models.skill import Skill, SkillSettings, SkillType
