+++ skill-manager/skill_manager/models/prediction.py
@@ -1,0 +1,35 @@
+from datetime import datetime
+from typing import List, Optional
+
+from pydantic import Field
+from square_skill_api.models.prediction import Prediction as SkillPrediction
+
+from skill_manager.mongo.mongo_model import MongoModel
+from skill_manager.mongo.py_object_id import PyObjectId
+
+
+class Prediction(MongoModel):
+    """Holds all prediction information, i.e. the output of a skill when queried."""
+
+    id: Optional[PyObjectId] = Field(
+        None, description="Identifier generated by mongoDB."
+    )
+    skill_id: PyObjectId = Field(
+        ..., description="Identifier of the skill that generated the prediction."
+    )
+    skill_name: str = Field(
+        ..., description="Name of the skill that generated the prediction."
+    )
+    query: str = Field(
+        ..., description="The input query that resulted in the prediction."
+    )
+    query_time: datetime = Field(
+        default_factory=datetime.now,
+        description="Time when the prediction was generated.",
+    )
+    user_id: str = Field(..., description="User that issued the query.")
+    predictions: List[SkillPrediction]
+    data_sets: List[str] = Field(
+        [],
+        description="All datasets with which the skill was trained",
+    )
