+++ skill-manager/skill_manager/models/skill.py
@@ -3,7 +3,6 @@ from enum import Enum
 from typing import Dict, List, Optional
 
 from pydantic import BaseModel, Field, validator
-from square_skill_api.models.prediction import Prediction as SkillPrediction
 
 from skill_manager.mongo.mongo_model import MongoModel
 from skill_manager.mongo.py_object_id import PyObjectId
@@ -129,30 +128,3 @@ class Skill(MongoModel):
                 "data_sets": ["HotpotQA", "SQuAD"],
             }
         }
-
-
-class Prediction(MongoModel):
-    """Holds all prediction information, i.e. the output of a skill when queried."""
-
-    id: Optional[PyObjectId] = Field(
-        None, description="Identifier generated by mongoDB."
-    )
-    skill_id: PyObjectId = Field(
-        ..., description="Identifier of the skill that generated the prediction."
-    )
-    skill_name: str = Field(
-        ..., description="Name of the skill that generated the prediction."
-    )
-    query: str = Field(
-        ..., description="The input query that resulted in the prediction."
-    )
-    query_time: datetime = Field(
-        default_factory=datetime.now,
-        description="Time when the prediction was generated.",
-    )
-    user_id: str = Field(..., description="User that issued the query.")
-    predictions: List[SkillPrediction]
-    data_sets: List[str] = Field(
-        [],
-        description="All datasets with which the skill was trained",
-    )
